git-svn-id: https://qfaktury.svn.sourceforge.net/svnroot/qfaktury/trunk@122 920ec645-f03f-0410-9fdc-e8aa689e005b

+++ Towary.h
@@ -14,7 +14,7 @@ public:
 public slots:
 	void okClick();
 	void spinChanged(int a);
-	void nettoChanged(int a);
+	void nettoChanged(double a);
 	void pkwiuGet();
 private:
 	QStringList netto;
